requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
bcrypt
email-validator
python-multipart
